feat(MobileMenuV3): WalletSelector: Renamed file... #476
