+++ src/components/WalletComponent/index.tsx
